tools/requirements.txt
requests>=2.28
Pillow>=9.0
rembg>=2.0
